Break down two components
