+++ frontend/src/metabase/admin/settings/components/widgets/GroupMappingsWidget/DeleteGroupMappingModal/DeleteGroupMappingModal.unit.spec.tsx
@@ -2,8 +2,6 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import type { GroupIds, DNType, ValueType } from "./DeleteGroupMappingModal";
-
 import DeleteGroupMappingModal, {
   DeleteGroupMappingModalProps,
 } from "./DeleteGroupMappingModal";
