+++ frontend/src/metabase/admin/settings/components/widgets/GroupMappingsWidget/LDAPMappingRow/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./LDAPMappingRow";
