+++ frontend/src/metabase/admin/settings/components/widgets/GroupMappingsWidget/types.ts
@@ -1,0 +1,2 @@
+export type GroupIds = number[];
+export type DNType = string;
